mejorando interacciones entre templates
