+++ blog_opiniones/templates/blog_opiniones/about.html
@@ -10,9 +10,10 @@
 {% endblock %}
 
 {% block contenido %}
-    Integrantes:
-    Mariano Ezequiel Schvartzman,
-    Mariano Moure Jorge.
+    
+Integrantes:<br>
+    Mariano Ezequiel Schvartzman,<br>
+    Mariano Moure Jorge.<br>
     Decidimos hacer un sistema de reservar y un blog de viajes ya que un miembro del grupo trabaja con un sistema similar en su día a día 
-    y quisimos ver como podíamos realizar un sistema propio
+    y quisimos ver como podíamos realizar un sistema propio experimentando las dificultades y complejidades que esto conlleva.
 {% endblock %}
