+++ blog_opiniones/templates/blog_opiniones/opinion_detail.html
@@ -21,7 +21,7 @@
     </div>
     <div class="container text-wrap py-1 my-1" style="word-break: break-all;">
         
-            {{object.opinion_detallada}}
+            {{object.opinion_detallada}}<br>
             {{object.firma}}
         
     </div>
