+++ blog_opiniones/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, include
 from blog_opiniones.views import *
 
 urlpatterns = [
