--- panel_reservas/templates/viajes/reserva_form.html
@@ -1,18 +1,0 @@
-{% extends 'viajes/base.html' %}
-{% load bootstrap4 %}
-
-{% block extrahead %}
-{{ form.media }}
-{% endblock %}
-
-{% block titulo%} Alta de nueva Reserva{% endblock %}
-{% block contenido %}
-
-    <form method="post">
-        {% csrf_token %}
-        {{form.as_p}}
-        <input type="submit" value="Agendar">
-    </form>
-<a href="{% url 'reserva-lista' %}">Volver a mis reservas</a>
-
-{% endblock %}
