--- panel_reservas/templates/viajes/reserva_list.html
@@ -1,17 +1,0 @@
-{% extends 'viajes/base.html' %}
-{% load bootstrap4 %}
-
-{% block titulo%} Lista de reservas {% endblock %}
-{% block contenido %}
-{% block buscar %} {% endblock %}
-<ul>
-{% for reserva in object_list %}
-<li> Nombre: {{reserva.nombre}}, Fecha: {{reserva.fecha_reserva}}, Hora: {{reserva.hora_reserva}}, Número de reserva: {{reserva.id}}
-    <a href="{% url 'reserva-borrar' reserva.id%}">Borrar</a>
-    <a href="{% url 'reserva-actualizar' reserva.id%}">Actualizar</a>
-</li>
-{% endfor %}    
-<a href="{% url 'reserva-crear'%}">Crear nueva reserva</a>
-</ul>
-
-{% endblock%}
