+++ viajes/templates/viajes/alta_reserva.html
@@ -16,5 +16,6 @@
   </form>
 
   <h2 style="color:green">{{msg_exito}}</h2>
+  <a href="{% url 'reservas' %}">Volver a mis reservas</a>
 
 {% endblock %}
