+++ viajes/templates/viajes/reservas.html
@@ -1,9 +1,43 @@
-{% extends 'viajes/base.html' %}
-{% block titulo%} Lista de reservas {% endblock %}
+{% extends 'blog_opiniones/base.html' %}
+
+{%block nav %}
+<!-- Responsive navbar-->
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+    <div class="container">
+        <a class="navbar-brand" href="{% url 'index-blog' %}">Mi Viaje!</a>
+        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><span
+                class="navbar-toggler-icon"></span></button>
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul class="navbar-nav ms-auto mb-2 mb-lg-0"></ul>
+                <li class="nav-item"><a class="nav-link" href="{% url 'index-blog' %}">Home</a></li>
+                <li class="nav-item"><a class="nav-link" href="{% url 'about' %}">About</a></li>
+                <li class="nav-item"><a class="nav-link active" aria-current="page" href="#">Blog</a></li>
+            </ul>
+        </div>
+    </div>
+</nav>
+{% endblock %}
+   <!-- Page header with logo and tagline-->
+   <header class="py-5 bg-light border-bottom mb-4">
+    {% block header %}
+    <div class="container">
+        <div class="text-center my-5">
+            <h1 class="fw-bolder">Mi Viaje!</h1>
+            <p class="lead mb-0">A viajar por el mundo!</p>
+        </div>
+    </div>
+    {% endblock %}
+</header>
 {% block contenido %}
   {% for reserva in lista_reservas %}
       <ul>
           <li> Nombre: {{reserva.nombre}}, Destino: {{reserva.destino_reserva}}, Fecha: {{reserva.fecha_reserva}}, Hora: {{reserva.hora_reserva}}, ReservaN°: {{reserva.id}}</li>
+          <a href="{% url 'reserva-borrar' reserva.id%}">Borrar</a>
+      </li>
+      {% endfor %}    
       </ul>
-  {% endfor %}
-{% endblock %}
+      <li><a href="{% url 'reservar'%}">Crear nueva reserva</a></li>
+      <li><a href="{% url 'index-blog' %}">Volver al blog</a></li>
+      {% endblock %}
+
