+++ vuelos/templates/vuelos/alta_vuelo.html
@@ -15,5 +15,6 @@
   </form>
 
   <h2 style="color:green">{{msg_exito}}</h2>
+  <a href="{% url 'vuelos' %}">Volver a mis reservas</a>
 
   {% endblock %}
