+++ vuelos/templates/vuelos/buscar_vuelo.html
@@ -1,5 +1,34 @@
-{% extends 'viajes/base.html' %}
-{% block titulo%} Buscar Vuelo {% endblock %}
+{% extends 'blog_opiniones/base.html' %}
+
+{%block nav %}
+<!-- Responsive navbar-->
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+    <div class="container">
+        <a class="navbar-brand" href="{% url 'index-blog' %}">Mi Viaje!</a>
+        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><span
+                class="navbar-toggler-icon"></span></button>
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul class="navbar-nav ms-auto mb-2 mb-lg-0"></ul>
+                <li class="nav-item"><a class="nav-link" href="{% url 'index-blog' %}">Home</a></li>
+                <li class="nav-item"><a class="nav-link" href="{% url 'about' %}">About</a></li>
+                <li class="nav-item"><a class="nav-link active" aria-current="page" href="#">Blog</a></li>
+            </ul>
+        </div>
+    </div>
+</nav>
+{% endblock %}
+   <!-- Page header with logo and tagline-->
+   <header class="py-5 bg-light border-bottom mb-4">
+    {% block header %}
+    <div class="container">
+        <div class="text-center my-5">
+            <h1 class="fw-bolder">Mi Viaje!</h1>
+            <p class="lead mb-0">Busca tu vuelo</p>
+        </div>
+    </div>
+    {% endblock %}
+</header>
 {% block contenido %}
   
   <form action="/buscar-vuelo" method="post">
